Moving openvino_backend.py from openvino.quantization to quantization module (#1745)
### Changes

Moving openvino_backend.py from openvino.quantization to quantization
module

### Reason for changes

Align with NNCF structure

### Related tickets

N/A

### Tests

N/A
