+++ nncf/quantization/algorithms/min_max/algorithm.py
@@ -215,7 +215,7 @@ class MinMaxQuantization(Algorithm):
 
             self._backend_entity = ONNXMinMaxAlgoBackend()
         elif model_backend == BackendType.OPENVINO:
-            from nncf.openvino.quantization.algorithms.min_max.openvino_backend import OVMinMaxAlgoBackend
+            from nncf.quantization.algorithms.min_max.openvino_backend import OVMinMaxAlgoBackend
 
             self._backend_entity = OVMinMaxAlgoBackend()
         elif model_backend == BackendType.TORCH:
